+++ src/crypto_investigator/constants/__init__.py
@@ -1,6 +1,6 @@
 """Stable application-wide constants."""
 
 APP_NAME = "ChainSherlock"
-VERSION = "0.1.1"
+VERSION = "0.1.2"
 
 __all__ = ["APP_NAME", "VERSION"]